check for CPU capabilities before using SIMD instructions
