+++ go.mod
@@ -1,3 +1,8 @@
 module github.com/segmentio/encoding
 
 go 1.14
+
+require (
+	github.com/klauspost/cpuid/v2 v2.0.5 // indirect
+	github.com/mmcloughlin/avo v0.0.0-20210104032911-599bdd1269f4 // indirect
+)
